# public/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Assinix",
        "version": "1.0.0",
        "description": "Assinix \u00e9 uma API REST profissional em PHP/Laravel para gerenciamento de assinaturas, vencimentos e gastos recorrentes.\n\nA API contempla autentica\u00e7\u00e3o com Laravel Sanctum, CRUD de categorias, formas de pagamento e assinaturas, c\u00e1lculo financeiro, dashboard, cota\u00e7\u00e3o de moedas, auditoria, logs estruturados e documenta\u00e7\u00e3o OpenAPI."
    },
    "servers": [
        {
            "url": "http://localhost:8080/api/v1",
            "description": "Local Docker"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/currencies/refresh": {
            "post": {
                "operationId": "currencies.refresh",
                "tags": [
                    "Currency"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshCurrencyRateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Cota\u00e7\u00e3o atualizada com sucesso."
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "rate": {
                                                    "type": "array",
                                                    "items": {}
                                                },
                                                "conversion_source": {
                                                    "type": "string"
                                                },
                                                "updated_subscriptions_count": {
                                                    "type": "integer"
                                                }
                                            },
                                            "required": [
                                                "rate",
                                                "conversion_source",
                                                "updated_subscriptions_count"
                                            ]
                                        },
                                        "errors": {
                                            "type": "null"
                                        },
                                        "meta": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data",
                                        "errors",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    }
                }
            }
        },
        "/currencies/convert": {
            "get": {
                "operationId": "currencies.convert",
                "tags": [
                    "Currency"
                ],
                "parameters": [
                    {
                        "name": "amount",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "maximum": 999999999.99
                        }
                    },
                    {
                        "name": "from_currency",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/CurrencyCode"
                        }
                    },
                    {
                        "name": "to_currency",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/CurrencyCode"
                        }
                    },
                    {
                        "name": "force_refresh",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Convers\u00e3o realizada com sucesso."
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "conversion": {
                                                    "type": "object",
                                                    "properties": {
                                                        "original_amount": {
                                                            "type": "string"
                                                        },
                                                        "converted_amount": {
                                                            "type": "string"
                                                        },
                                                        "from_currency": {
                                                            "type": "string"
                                                        },
                                                        "to_currency": {
                                                            "type": "string"
                                                        },
                                                        "rate": {
                                                            "type": "string"
                                                        },
                                                        "conversion_source": {
                                                            "type": "string"
                                                        },
                                                        "quoted_at": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "required": [
                                                        "original_amount",
                                                        "converted_amount",
                                                        "from_currency",
                                                        "to_currency",
                                                        "rate",
                                                        "conversion_source",
                                                        "quoted_at"
                                                    ]
                                                }
                                            },
                                            "required": [
                                                "conversion"
                                            ]
                                        },
                                        "errors": {
                                            "type": "null"
                                        },
                                        "meta": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data",
                                        "errors",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "CurrencyCode": {
                "type": "string",
                "enum": [
                    "BRL",
                    "USD",
                    "EUR",
                    "GBP",
                    "ARS",
                    "CAD"
                ],
                "title": "CurrencyCode"
            },
...
            "RefreshCurrencyRateRequest": {
                "type": "object",
                "properties": {
                    "currency": {
                        "$ref": "#/components/schemas/CurrencyCode"
                    },
                    "update_subscriptions": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "currency"
                ],
                "title": "RefreshCurrencyRateRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "AuthorizationException": {
                "description": "Authorization error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}